Refactoring tbl_indexes.lib.php & tbl_chart.lib.php
